refactor: refactor fs throw Error

+++ src/fs/copy.js
@@ -1,4 +1,4 @@
-import {readdir, copyFile, mkdir} from 'fs/promises'
+import {readdir, copyFile, mkdir, access} from 'fs/promises'
 import {join, dirname} from 'path'
 import {fileURLToPath} from 'url'
 
@@ -10,16 +10,28 @@ export const copy = async () => {
   const destination = join(__dirname, 'files-copy');
 
   try {
+    if (await isExists(destination) || !await isExists(source)) {
+      throw new Error('FS operation failed');
+    }
     await mkdir(destination);
     const files = await readdir(source);
     for (let file of files) {
-      const sourceFile = path.join(source, file);
-      const destinationFile = path.join(destination, file);
+      const sourceFile = join(source, file);
+      const destinationFile = join(destination, file);
       await copyFile(sourceFile, destinationFile);
     }
   } catch (err) {
-    throw new Error('FS operation failed');
+    console.error(err);
   }
 };
+
+async function isExists(path) {
+  try {
+    await access(path)
+    return true;
+  } catch {
+    return false;
+  }
+}
 
 copy();
